+++ src/svs/inacademia_frontend.py
@@ -239,8 +239,18 @@ class InAcademiaFrontend(OpenIDConnectFrontend):
         return url_map
 
     def handle_backend_error(self, exception):
-        if "Authentication failed" in exception.message:
+        if 'Authentication failed' in exception.message:
             exception._message = ErrorDescription.AUTHENTICATION_ERROR_FROM_IDP[ERROR_DESC]
+            resp_rp = 'unknown'
+
+            if exception.state and exception.state[self.name] and exception.state[self.name]["oidc_request"]:
+                auth_req = self._get_authn_request_from_state(exception.state)
+
+                if auth_req and auth_req.get('client_id'):
+                    resp_rp = auth_req.get('client_id')
+
+            transaction_log(exception.state, self.config.get("response_exit_order", 600),
+                            "inacademia_frontend", "response", "exit", "failed", resp_rp, '',
+                            ErrorDescription.AUTHENTICATION_ERROR_FROM_IDP[LOG_MSG])
 
         return super().handle_backend_error(exception)
-